+++ test/unit/component/parser/json/BpmnJsonParser.lane.test.ts
@@ -145,44 +145,6 @@ describe('parse bpmn as json for lane', () => {
     expect(warning.flowNodeRef).toBe('event_id_0');
   });
 
-  it('json containing one process with a single lane with flowNodeRef as object & flowNode not parsed', () => {
-    const json = {
-      definitions: {
-        targetNamespace: '',
-        process: {
-          lane: { id: 'Lane_12u5n6x', flowNodeRef: 'event_id_0' },
-        },
-        BPMNDiagram: {
-          BPMNPlane: {
-            BPMNShape: {
-              id: 'Lane_1h5yeu4_di',
-              bpmnElement: 'Lane_12u5n6x',
-              isHorizontal: true,
-              Bounds: { x: 362, y: 232, width: 36, height: 45 },
-            },
-          },
-        },
-      },
-    };
-
-    const model = parseJson(json);
-
-    expect(model.lanes).toHaveLength(1);
-    verifyShape(model.lanes[0], {
-      shapeId: 'Lane_1h5yeu4_di',
-      bpmnElementId: 'Lane_12u5n6x',
-      bpmnElementName: undefined,
-      bpmnElementKind: ShapeBpmnElementKind.LANE,
-      bounds: {
-        x: 362,
-        y: 232,
-        width: 36,
-        height: 45,
-      },
-      isHorizontal: true,
-    });
-  });
-
   it('json containing one process with a single lane with flowNodeRef as array', () => {
     const json = {
       definitions: {
